+++ lib/native/affise_api_method.dart
@@ -23,8 +23,10 @@ enum AffiseApiMethod {
   IS_FIRST_RUN,
 
   // callbacks
-  GET_REFERRER_CALLBACK,
-  GET_REFERRER_VALUE_CALLBACK,
+  GET_REFERRER_URL_CALLBACK,
+  GET_REFERRER_URL_VALUE_CALLBACK,
+  GET_REFERRER_ON_SERVER_CALLBACK,
+  GET_REFERRER_ON_SERVER_VALUE_CALLBACK,
   REGISTER_DEEPLINK_CALLBACK,
   SKAD_REGISTER_ERROR_CALLBACK,
   SKAD_POSTBACK_ERROR_CALLBACK,
@@ -99,10 +101,14 @@ extension AffiseApiMethodExt on AffiseApiMethod {
         return "get_providers";
       case AffiseApiMethod.IS_FIRST_RUN:
         return "is_first_run";
-      case AffiseApiMethod.GET_REFERRER_CALLBACK:
-        return "get_referrer_callback";
-      case AffiseApiMethod.GET_REFERRER_VALUE_CALLBACK:
-        return "get_referrer_value_callback";
+      case AffiseApiMethod.GET_REFERRER_URL_CALLBACK:
+        return "get_referrer_url_callback";
+      case AffiseApiMethod.GET_REFERRER_URL_VALUE_CALLBACK:
+        return "get_referrer_url_value_callback";
+      case AffiseApiMethod.GET_REFERRER_ON_SERVER_CALLBACK:
+        return "get_referrer_on_server_callback";
+      case AffiseApiMethod.GET_REFERRER_ON_SERVER_VALUE_CALLBACK:
+        return "get_referrer_on_server_value_callback";
       case AffiseApiMethod.REGISTER_DEEPLINK_CALLBACK:
         return "register_deeplink_callback";
       case AffiseApiMethod.SKAD_REGISTER_ERROR_CALLBACK: