+++ lib/native_api_interface.dart
@@ -24,6 +24,10 @@ abstract class AffiseIOSApi {
   void registerAppForAdNetworkAttribution(ErrorCallback completionHandler);
 
   void updatePostbackConversionValue(int fineValue, CoarseValue coarseValue, ErrorCallback completionHandler);
+
+  void getReferrerOnServer(ReferrerCallback callback);
+
+  void getReferrerOnServerValue(ReferrerKey key, ReferrerCallback callback);
 }
 
 abstract class AffiseDebug {